+++ Tevling/Services/IDevService.cs
@@ -1,0 +1,11 @@
+using Tevling.Strava;
+
+namespace Tevling.Services;
+
+public interface IDevService
+{
+    public void SetRandomEnabled(bool changed);
+    public bool IsRandomEnabled();
+    
+    public DetailedActivity GetActivity(long stravaId);
+}
