--- Tevling/Services/IRandomToggleService.cs
@@ -1,7 +1,0 @@
-namespace Tevling.Services;
-
-public interface IRandomToggleService
-{
-    public void SetRandomEnabled(bool changed);
-    public bool IsEnabled();
-}
